# auth_service/docs/swagger.yaml
basePath: /
definitions:
...
  internal_http_server_handlers_oauth_callback.Response:
    properties:
      access_token:
        example: fkajeDJ1p3FJ...
        type: string
      error:
        example: error
        type: string
      refresh_token:
        example: abcDEF123...
        type: string
      status:
        example: ok
        type: string
    type: object
  internal_http_server_handlers_oauth_link.Response:
    properties:
      error:
        example: error
        type: string
      redirect_url:
        example: https://www.google.com/
        type: string
      status:
        example: ok
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: Сервис авторизации
  title: Auth Service API
  version: "1.0"
paths:
...
  /auth/oauth/{provider}/callback:
    get:
      description: |-
        Обрабатывает перенаправление от OAuth-провайдера после того, как пользователь предоставил или отклонил доступ.
        Проверяет параметр state на соответствие ранее сохранённому значению (авторизация или привязка аккаунта),
        обменивает код авторизации на токены провайдера и в зависимости от типа операции
        либо выдает новую пару access и refresh токенов,
        либо привязывает OAuth-провайдера к существующему аккаунту.
      parameters:
      - description: 'Название OAuth-провайдера (например: google, github)'
        in: path
        name: provider
        required: true
        type: string
      - description: Код авторизации, полученный от OAuth-провайдера
        in: query
        name: code
        required: true
        type: string
      - description: Токен состояния (state), должен совпадать со значением, выданным
          при начале авторизации или привязки аккаунта
        in: query
        name: state
        required: true
        type: string
      - description: Код ошибки, возвращаемый OAuth-провайдером, если пользователь
          отказал в доступе
        in: query
        name: error
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация или привязка аккаунта
          schema:
            $ref: '#/definitions/internal_http_server_handlers_oauth_callback.Response'
        "400":
          description: Пользователь отказал в доступе, отсутствуют параметры code/state,
            указан некорректный app_id либо state недействителен или истёк
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "403":
          description: Email, полученный от OAuth-провайдера, не подтверждён
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Указанный OAuth-провайдер не поддерживается
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "409":
          description: 'Конфликт данных: аккаунт с таким email уже существует либо
            OAuth-провайдер уже привязан к другому аккаунту'
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Callback OAuth-провайдера
      tags:
      - oauth
  /auth/oauth/{provider}/link:
    get:
      description: |-
        Запускает OAuth2-процесс привязки стороннего OAuth-провайдера к аккаунту
        текущего аутентифицированного пользователя.
        Требует действительный access-токен. Идентификатор пользователя
        извлекается из его данных, а не из параметров запроса,
        поэтому привязка всегда выполняется для текущей сессии.
      parameters:
      - description: 'Название OAuth-провайдера (например: google, github)'
        in: path
        name: provider
        required: true
        type: string
      - description: URL, на который будет выполнено перенаправление после завершения
          авторизации. Должен входить в список разрешённых адресов.
        in: query
        name: redirect_uri
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ссылка для перехода к OAuth-провайдеру
          schema:
            $ref: '#/definitions/internal_http_server_handlers_oauth_link.Response'
        "400":
          description: Некорректный redirect_uri или app_id
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Access-токен отсутствует, недействителен или истёк
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "404":
          description: OAuth-провайдер не поддерживается
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Привязка OAuth-провайдера к существующему аккаунту
      tags:
      - oauth
...
  /auth/password/forgot:
    post:
      consumes:
      - application/json
      description: |-
        Запускает процесс сброса пароля для указанного адреса электронной почты.
        Независимо от того, существует ли аккаунт с указанным email,
        всегда возвращает ответ с кодом 200, чтобы исключить возможность
        определения существования аккаунтов.
        Если аккаунт существует, на указанный email будет отправлено
        письмо со ссылкой для сброса пароля. Ошибки отправки письма
        фиксируются на стороне сервера и не влияют на ответ API.
      parameters:
      - description: Адрес электронной почты пользователя
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Запрос успешно принят
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Некорректное тело запроса или ошибка валидации
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "429":
          description: Превышен допустимый лимит запросов
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Запрос на сброс пароля
      tags:
      - auth
...
swagger: "2.0"